webpage/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
Django==5.1
djangorestframework==3.15.2
drf-yasg==1.21.7
idna==3.8
inflection==0.5.1
packaging==24.1
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
setuptools==74.1.1
sqlparse==0.5.1
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
gunicorn==21.2.0
